+++ src/CronFieldValidator.php
@@ -18,12 +18,19 @@ interface CronFieldValidator
 
     /**
      * When a CRON expression is not satisfied, this method is used to increment
-     * or decrement a DateTime or a DateTimeImmutable object by the unit of the cron field.
+     * a DateTime or a DateTimeImmutable object by the unit of the cron field.
      *
-     * @param bool $invert Set to TRUE to decrement
-     * @param string|null $fieldExpression Set parts to use
+     * @param string|null $fieldExpression the field expression value if available
      */
-    public function increment(DateTime|DateTimeImmutable $date, bool $invert = false, string|null $fieldExpression = null): DateTime|DateTimeImmutable;
+    public function increment(DateTime|DateTimeImmutable $date, string|null $fieldExpression = null): DateTime|DateTimeImmutable;
+
+    /**
+     * When a CRON expression is not satisfied, this method is used to decrement
+     * a DateTime or a DateTimeImmutable object by the unit of the cron field.
+     *
+     * @param string|null $fieldExpression the field expression value if available
+     */
+    public function decrement(DateTime|DateTimeImmutable $date, string|null $fieldExpression = null): DateTime|DateTimeImmutable;
 
     /**
      * Validates a CRON expression for a given field.