+++ src/DayOfMonthValidator.php
@@ -78,16 +78,18 @@ final class DayOfMonthValidator extends FieldValidator
         };
     }
 
-    public function increment(DateTime|DateTimeImmutable $date, bool $invert = false, string $fieldExpression = null): DateTime|DateTimeImmutable
+    public function increment(DateTime|DateTimeImmutable $date, string $fieldExpression = null): DateTime|DateTimeImmutable
     {
-        return match ($invert) {
-            true => $date
-                ->setDate((int) $date->format('Y'), (int) $date->format('n'), (int) $date->format('j') - 1)
-                ->setTime(23, 59),
-            default => $date
-                ->setDate((int) $date->format('Y'), (int) $date->format('n'), (int) $date->format('j') + 1)
-                ->setTime(0, 0),
-        };
+        return $date
+            ->setDate((int) $date->format('Y'), (int) $date->format('n'), (int) $date->format('j') + 1)
+            ->setTime(0, 0);
+    }
+
+    public function decrement(DateTime|DateTimeImmutable $date, string $fieldExpression = null): DateTime|DateTimeImmutable
+    {
+        return $date
+            ->setDate((int) $date->format('Y'), (int) $date->format('n'), (int) $date->format('j') - 1)
+            ->setTime(23, 59);
     }
 
     /**