+++ src/DayOfWeekValidatorTest.php
@@ -34,7 +34,7 @@ final class DayOfWeekValidatorTest extends TestCase
         $f = new DayOfWeekValidator();
         self::assertSame('2011-03-16 00:00:00', $f->increment(new DateTime('2011-03-15 11:15:00'))->format('Y-m-d H:i:s'));
 
-        self::assertSame('2011-03-14 23:59:00', $f->increment(new DateTime('2011-03-15 11:15:00'), true)->format('Y-m-d H:i:s'));
+        self::assertSame('2011-03-14 23:59:00', $f->decrement(new DateTime('2011-03-15 11:15:00'))->format('Y-m-d H:i:s'));
     }
 
     public function testIncrementsDateImmutable(): void
@@ -42,7 +42,7 @@ final class DayOfWeekValidatorTest extends TestCase
         $f = new DayOfWeekValidator();
         self::assertSame('2011-03-16 00:00:00', $f->increment(new DateTimeImmutable('2011-03-15 11:15:00'))->format('Y-m-d H:i:s'));
 
-        self::assertSame('2011-03-14 23:59:00', $f->increment(new DateTimeImmutable('2011-03-15 11:15:00'), true)->format('Y-m-d H:i:s'));
+        self::assertSame('2011-03-14 23:59:00', $f->decrement(new DateTimeImmutable('2011-03-15 11:15:00'))->format('Y-m-d H:i:s'));
     }
 
     /**