+++ src/HourValidator.php
@@ -23,7 +23,7 @@ final class HourValidator extends FieldValidator
         return $this->isSatisfied((int) $date->format('H'), $fieldExpression);
     }
 
-    public function increment(DateTime|DateTimeImmutable $date, bool $invert = false, string $fieldExpression = null): DateTime|DateTimeImmutable
+    public function increment(DateTime|DateTimeImmutable $date, string $fieldExpression = null): DateTime|DateTimeImmutable
     {
         // Change timezone to UTC temporarily. This will
         // allow us to go back or forwards and hour even
@@ -32,12 +32,6 @@ final class HourValidator extends FieldValidator
             $interval = new DateInterval('PT1H');
             $timezone = $date->getTimezone();
             $date = $date->setTimezone(new DateTimeZone('UTC'));
-            if ($invert) {
-                $date = $date->sub($interval)->setTimezone($timezone);
-
-                return $date->setTime((int) $date->format('H'), 59);
-            }
-
             $date = $date->add($interval)->setTimezone($timezone);
 
             return $date->setTime((int) $date->format('H'), 0);
@@ -49,17 +43,39 @@ final class HourValidator extends FieldValidator
             []
         );
 
-        $hour = $hours[$this->computePosition((int) $date->format('H'), $hours, $invert)];
-        if ((!$invert && $date->format('H') >= $hour) || ($invert && $date->format('H') <= $hour)) {
-            return match ($invert) {
-                true => $date->sub(new DateInterval('P1D'))->setTime(23, 59),
-                default => $date->add(new DateInterval('P1D'))->setTime(0, 0),
-            };
+        $hour = $hours[$this->computePosition((int) $date->format('H'), $hours, false)];
+        if ($date->format('H') >= $hour) {
+            return $date->add(new DateInterval('P1D'))->setTime(0, 0);
         }
 
-        return match ($invert) {
-            true => $date->setTime($hour, 59),
-            default => $date->setTime($hour, 0),
-        };
+        return $date->setTime($hour, 0);
+    }
+
+    public function decrement(DateTime|DateTimeImmutable $date, string $fieldExpression = null): DateTime|DateTimeImmutable
+    {
+        // Change timezone to UTC temporarily. This will
+        // allow us to go back or forwards and hour even
+        // if DST will be changed between the hours.
+        if (null === $fieldExpression || $fieldExpression == '*') {
+            $interval = new DateInterval('PT1H');
+            $timezone = $date->getTimezone();
+            $date = $date->setTimezone(new DateTimeZone('UTC'));
+            $date = $date->sub($interval)->setTimezone($timezone);
+
+            return $date->setTime((int) $date->format('H'), 59);
+        }
+
+        $hours = array_reduce(
+            str_contains($fieldExpression, ',') ? explode(',', $fieldExpression) : [$fieldExpression],
+            fn (array $hours, string $part): array => array_merge($hours, $this->getRangeForExpression($part, 23)),
+            []
+        );
+
+        $hour = $hours[$this->computePosition((int) $date->format('H'), $hours, true)];
+        if ($date->format('H') <= $hour) {
+            return $date->sub(new DateInterval('P1D'))->setTime(23, 59);
+        }
+
+        return $date->setTime($hour, 59);
     }
 }